Major Changes: Version 2.0
• All the hazards now have a special parameter called hazardID that
determines its placement on the screen. Simply assign each hazard a
different integer ID and it will place it accordingly. Each increment of
the hazardID shifts the hazard 50 pixels to the right.
• Collisions are now turned off when the hazards are in the menu bar.
They activate again when the mouse gets within a 5 pixel distance,
allowing player2 to pick up the hazards. We can now build levels as high
as we want without worrying about the hazards bumping into the
platforms.
• Increased the minimum release radius to a distance of about 25. So if
a player 2 tries to release an object to close to the player 1, it
offsets it and drops it just behind the player every time.
• All of the hazards now have a respawn timer that can be set to
anything above 0. So if the respawn timer for a skeleton to 5 for
example, after the player 2 releases the skeleton, it will respawn back
to its original position after 5 seconds. If you don’t want a hazard to
respawn simply set the timer to something high ( eg 10e8).
• Changed the control freak to only affect the ‘A’ and ‘D’ keys while
leaving the shift and space keys alone as mixing up all the controls
makes it too hard for player 1
• Player 1 now has an attribute called controlFreakTimer which sets the
effective time until the control freak wears out and controls are reset
again. Set to an integer above 0 (in seconds).
• Stopped the player from moving after a death or a win (froze his
velocity)
• Cleaned up a little of the code in general